+++ src/js/components/slideshow.js
@@ -85,8 +85,10 @@
 
                                 })
                                 .attr('src', [src, (src.indexOf('?') > -1 ? '&':'?'), 'enablejsapi=1&api=1'].join(''))
-                                .addClass('uk-position-absolute')
-                                .css('pointer-events', 'none');
+                                .addClass('uk-position-absolute');
+
+                                // disable pointer events
+                                if(!UI.support.touch) media.css('pointer-events', 'none');
 
                             placeholder = true;