+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### WIP
   - Added slidenav-inverted modifier
+  - Added prevent registering a component multiple times with the same name
   - Added direction info to the uk-scroll event
   - Added $.UIkit.Utils.str2json helper function
   - Updated active paremeter for the switcher component to allow false as value
+  - Disable pointer events none on iframes in slideshows on touch devices
   - Fixed automute in slideshow component
   - Fixed hidden text in dotnav
   - Fixed prevent editing form fields within a sortable component